backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Cython==3.0.8
Deprecated==1.2.18
dnspython==2.7.0
docopt==0.6.2
entropy==0.1.5
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Limiter==3.12
flatbuffers==24.3.25
gast==0.6.0
gdcm==1.1
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
importlib_resources==6.4.5
itsdangerous==2.2.0
jax==0.4.26
Jinja2==3.1.6
joblib==1.5.1
kaggle==1.6.3
kagglehub==0.3.3
keras==3.6.0
lazy_loader==0.3
libclang==18.1.1
limits==5.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.4.1
mne==1.6.1
namex==0.0.8
nibabel==5.3.0
nolds==0.5.2
numpy==2.3.0
opencv-contrib-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pmdarima==2.0.4
pooch==1.8.0
protobuf==3.20.3
pydicom==3.0.0
Pygments==2.19.1
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.2.0
pylibjpeg-openjpeg==2.3.0
pymongo==4.13.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
rich==13.9.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sounddevice==0.4.6
soundfile==0.12.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
wfdb==4.1.2
wrapt==1.17.2
xgboost==2.1.1
